2019/1/1联调联试修改

+++ main.py
@@ -35,7 +35,7 @@ while True:
         angle = 35 / (10.1 - left_length) # 角度
         turns = angle / 3.18 # 圈数
         # 手在，出纸
-        stepper.forward(3 / 1000.0, turns * 50)
+        stepper.forward(3 / 1000.0, int(turns * 50))
         lock = True
     if  GPIO.input(infrared_channel) == 0 :
         lock = False
